Add magic number definitions for Kano applications
KanoComputing/peldins#1511
Add definitions for Make Minecraft, Make Pong and Make Snake to the
magic number definition file. The scheme is that `KANO` is included in
the first few bytes of the file followed by the level name in upper
case.

For Make Minecraft, for example, including at the top of the file the
line:

<!--KANO_MAKE-MINECRAFT-->

will associate the file with the Minecraft MIME type.