kano-uixinit: Speed up the wallpaper screen
The wallpaper settings screen will be launched in the background while
the video plays so it's ready for the user when the playback finishes.

Tested on the Pi and it seems to work fine. The window is waiting there
when the video is over. I didn't notice any stutters/performance hit
of doing this in the video. I guess that's because it's being done on
the GPU.