Support for a USB microphone dongle
The USB microphone I worked with seemed to be recognised as C-Media
CM108. The setup before defined a master channel for controlling
the volume in steps. Here I've defined the capture channel on the
sound card to come by default from a hardware "sound card" with
index 1, which will be the USB mic. There is a chance this will
create problems due to the index when plugging in other audio
devices.