Removed local installation of the ALSA config file
Previously, the ALSA config file was installed both under /etc/
and /etc/skel. The latter meant that the config was installed for
every user under ~/.asoundrc, however, no changes are expected to
be in user space. The reason for this was a need to change some
options system wide (max_dB, see kano-settings).