Remove the Volume widget in LXPanel in Classic Mode
This is necessary because the volume plugin has never heard of the
system ALSA config - /etc/asound.conf and when it doesn't find one
in the user homedir, it creates one from scratch. This breaks our
volume levels and capture channel for the USB microphone. Removing
it until we find a solution to the problem.