Fixed X Screen Saver starting on a new user first boot
The issue was that the code for disabling the X Screen Saver was
below the point where the script exited quickly when the Overture
onboarding was running. Moving these lines above will ensure that
the X Screen Saver does not start before ours blanking the screen
entirely on the first boot.